Qemotions Game final version
Qemotions is an educational and engaging game that combines quantum computing concepts with puzzle-solving gameplay. By completing levels and optimizing circuits, players will gain a better understanding of quantum mechanics while having fun. The game's gradual complexity and introduction of new quantum properties ensure an enriching experience for players of all skill levels.

this is a video on how to play it :
https://vimeo.com/865266521

and this is the link for playing it through the browser : 
https://mohamed-idriss.itch.io/qmetion